fix(homePageNumbers): use model when query is simple
The previous system, using Services, was returning string instead of a number for the "count" methods. It generated a warning on the client side ('got string instead of number').

For requests that simple, we can directly use the Sails model. Also, they directly return a number when using their count() method.