improvement(authRendering): add isUser method to AuthHelper
isUser() != isAuth()
You can be auth as an admin but if you are not an user in the database, you can't perform some actions (like submitting a document).
That's why we need to make a distinction between someone authenticated (an user, a moderator or an admin) and someone who is an user.